+++ changelog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added info on control and process variables in the PID tuner tool, including ability to change max/min limits for both.
 - Control variable value is rounded to precision before being displayed.
 - Added more information to the info collapsible in the PID tuner tool.
+- Added input UI elements to set the automatic PV mode step values in the PID tuner tool, closes #174.
 
 ## [v3.4.2] - 2018-04-27