+++ changelog.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [v3.4.3] - 2018-05-02
+
 ### Added
 - Added info on control and process variables in the PID tuner tool, including ability to change max/min limits for both.
 - Control variable value is rounded to precision before being displayed.